# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Proyectos",
        "contact": {
            "email": "andy.ancajima@mimco.com.pe"
        },
        "version": "1"
    },
    "paths": {
...
        "/GestionProyectos/public/index.php/regi_moti": {
            "post": {
                "tags": [
                    "Gestion Motivo"
                ],
                "summary": "Registras nuevos  motivos",
                "operationId": "App\\Http\\Controllers\\MotivoController::regi_moti",
                "parameters": [
                    {
                        "name": "intIdTipoMoti",
                        "in": "path",
                        "description": "Ingrese el tipo de motivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2"
                    },
                    {
                        "name": "varDescripcion",
                        "in": "path",
                        "description": "Ingrese la descripcion del motivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "RETRASO"
                    },
                    {
                        "name": "intIdEsta",
                        "in": "path",
                        "description": "Ingrese el estado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3"
                    },
                    {
                        "name": "acti_usua",
                        "in": "path",
                        "description": "Ingresar el usuario que va registrar un nuevo motivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "andy_ancajima"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdTipoMoti": {
                                        "type": "string"
                                    },
                                    "varDescripcion": {
                                        "type": "string"
                                    },
                                    "intIdEsta": {
                                        "type": "string"
                                    },
                                    "acti_usua": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdTipoMoti": "2",
                                    "varDescripcion": "RETRASO",
                                    "intIdEsta": "3",
                                    "acti_usua": "andy_ancajima"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro Satisfactorio."
                    },
                    "407": {
                        "description": "La descripcion ya existe."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionProyectos/public/index.php/actu_moti": {
            "post": {
                "tags": [
                    "Gestion Motivo"
                ],
                "summary": "Actualizar un motivo",
                "operationId": "App\\Http\\Controllers\\MotivoController::actu_moti",
                "parameters": [
                    {
                        "name": "intIdMoti",
                        "in": "path",
                        "description": "Ingrese el id de motivo que se va actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2"
                    },
                    {
                        "name": "intIdTipoMoti",
                        "in": "path",
                        "description": "Ingrese el id del tipo motivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2"
                    },
                    {
                        "name": "varDescripcion",
                        "in": "path",
                        "description": "Ingrese La descripcion del motivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "PRUEBA_FALTA"
                    },
                    {
                        "name": "intIdEsta",
                        "in": "path",
                        "description": "Ingrese el id estado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3"
                    },
                    {
                        "name": "usua_modi",
                        "in": "path",
                        "description": "Ingrese el nombre del usuario quien va modificar un motivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "jose_castillo"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdTipoMoti": {
                                        "type": "string"
                                    },
                                    "varDescripcion": {
                                        "type": "string"
                                    },
                                    "intIdEsta": {
                                        "type": "string"
                                    },
                                    "usua_modi": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdMoti": "2",
                                    "intIdTipoMoti": "2",
                                    "varDescripcion": "PRUEBA_FALTA",
                                    "intIdEsta": "3",
                                    "usua_modi": "jose_castillo"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizacion Satisfactoria"
                    },
                    "407": {
                        "description": "La Descripcion ya existe."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
        "/GestionProyectos/public/index.php/elim_mot": {
            "post": {
                "tags": [
                    "Gestion Motivo"
                ],
                "summary": "obtiene datos del usuario a través del dni",
                "operationId": "App\\Http\\Controllers\\MotivoController::elim_mot",
                "parameters": [
                    {
                        "name": "intIdMoti",
                        "in": "path",
                        "description": "ingresar el id del motivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2"
                    },
                    {
                        "name": "usua_modi",
                        "in": "path",
                        "description": "ingresar el usuario que ha modificado un motivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "omar_chacon"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdMoti": {
                                        "type": "string"
                                    },
                                    "usua_modi": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdMoti": "2",
                                    "usua_modi": "omar_chacon"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se ha eliminado."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
    }
}